brain-sds-memory/requirements.txt
grpcio==1.23.0
grpcio-tools==1.23.0
protobuf==3.9.1
six==1.12.0
